+++ client/Packages/com.beamable/Runtime/Modules/Purchasing/Steam/SteamPurchasingModule.cs
@@ -1,4 +1,7 @@
-﻿using Beamable.Common.Dependencies;
+﻿// Unity IAP v5 removed AbstractPurchasingModule / IStoreConfiguration; the v5 Steam store is
+// registered through UnityIAPServices in UnityBeamablePurchaser instead of this module.
+#if !UNITY_PURCHASING_5_OR_NEWER
+using Beamable.Common.Dependencies;
 using UnityEngine.Purchasing.Extension;
 
 namespace Beamable.Purchasing.Steam
@@ -17,4 +20,5 @@ namespace Beamable.Purchasing.Steam
 		}
 	}
 }
+#endif